Addressing legitimate warning messages #2
Fixed -Wall -Wextra warnings on UNIX build. 

This is big update. It allows to have keep all warnings turned on in build environment, and catch coding errors at earlier stages.

+++ wcm/eloadsave.cpp
@@ -279,7 +279,7 @@ void OperSaveFileThread::Run()
 
 			if ( count <= 0 ) { break; }
 
-			if ( count > sizeof( buf ) )
+			if ( count > (int)sizeof( buf ) )
 			{
 				count = sizeof( buf );
 			}
